Move cliclick.py into functions folder
