+++ trader.py
@@ -14,7 +14,7 @@ import datetime
 
 # Custom imports
 import config
-from cliclick import Cliclick
+from functions.cliclick import Cliclick
 import functions.utils as utils
 
 # Access the coordinates from the active system
